+++ XMedianY.java
@@ -9,13 +9,13 @@ class XMedianY extends MedianStrategy {
 
         // get a set of all timestamps;
         List<Place> places = this.allPlaces(trajectories);
-        List<Integer> timestamps = new LinkedList<Integer>();
+        List<Long> timestamps = new LinkedList<Long>();
         for (Place p : places) {
             if ( !timestamps.contains(p.getT()) ) timestamps.add(p.getT());
         }
 
         // for all timestamps, find the corresponding median Place
-        for (Integer t : timestamps) {
+        for (Long t : timestamps) {
             Place xMedian = this.findXMedianPlaceAtTime(trajectories, t);
             
             Place p = new Place(xMedian.getX(), xMedian.getY(), t);
